refactor(datagrid-web): rename directory
